backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
httpx==0.27.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydub==0.25.1
parselmouth==0.4.4
numpy==1.26.4
scipy==1.14.1
aiofiles==24.1.0
